Extract k8s CPU/memory parsers from kubectl.py to k8s_types.py
Move parse_k8s_cpu and parse_k8s_memory to k8s_types as public functions.
kubectl.py re-imports them as private aliases for backward compat.
Remove duplicate top_pod tests (covered by K8sServiceImpl tests) and
rename test file to test_k8s_parsers.py.
